fix: cast predictor input to float32 for mixed-precision backbones feat: add multi-seed benchmark results (seeds 123, 456)
- Fix BFloat16/Float32 dtype mismatch in predictor for Qwen backbone
- Seed 42: 100% ID, 0% OOD
- Seed 123: 100% ID, 14% OOD (partial generalization)
- Seed 456: 10% ID (training instability)

+++ src/model.py
@@ -129,7 +129,7 @@ class JEPAPredictor(nn.Module):
         Returns:
             Predicted next state embedding (B, D_s).
         """
-        combined = torch.cat([state, action], dim=-1)  # (B, D_s + D_a)
+        combined = torch.cat([state, action], dim=-1).float()  # (B, D_s + D_a)
         h = self.proj_in(combined).unsqueeze(1)  # (B, 1, hidden)
         h = self.transformer(h)  # (B, 1, hidden)
         return self.proj_out(h.squeeze(1))  # (B, D_s)
